[update] version 1.12.5 抽奖业务实践 积分扣除实践

+++ src/main/java/fun/gengzi/codecopy/business/luckdraw/entity/TbPrize.java
@@ -31,7 +31,7 @@ public class TbPrize {
     private Integer prizeNum;
     //概率
     @Column(name = "probability")
-    private BigDecimal probability;
+    private Double probability;
     //活动id
     @Column(name = "activityid")
     private String activityid;
